Correcciones en las clases Notas! * Se cambio a int la variable de frecuencia para requerir menor uso de memoria * Debido a que se requirio cambiar los valores de frecuencias, corregui los test * Ahora corren el 100% los test

+++ trunk/AlgoHero/src/algo3c1g2/modelo/nota/Do.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class Do extends Nota {
 
 	//Frecuencia caracteristica del tono Do en la octava 3
-	private final static float FRECUENCIA_DO = 60;
+	private final static int FRECUENCIA_DO = 60;
 	private final static int CUERDA = 1;
 	
 	public Do(){
