+++ trunk/AlgoHero/src/algo3c1g2/modelo/nota/Re.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class Re extends Nota {
 
 	//Frecuencia caracteristica del tono Re en la octava 3
-	private final static float FRECUENCIA_RE = 62;
+	private final static int FRECUENCIA_RE = 62;
 	private final static int CUERDA = 2;
 	
 	public Re(){
